# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface CompileRequest {
  /** Rust-inspired circuit definition source code */
  source: string;
}

export type CompileErrorSeverity =
  (typeof CompileErrorSeverity)[keyof typeof CompileErrorSeverity];

export const CompileErrorSeverity = {
  error: "error",
  fatal: "fatal",
} as const;

export interface CompileError {
  line: number;
  column: number;
  message: string;
  errorCode: string;
  severity: CompileErrorSeverity;
}

export interface CompileWarning {
  line: number;
  column: number;
  message: string;
  warningCode: string;
}

export type NetlistComponentCategory =
  (typeof NetlistComponentCategory)[keyof typeof NetlistComponentCategory];

export const NetlistComponentCategory = {
  passive: "passive",
  active_discrete: "active_discrete",
  active_ic: "active_ic",
  sensor: "sensor",
  module: "module",
  power: "power",
} as const;

export type PinType = (typeof PinType)[keyof typeof PinType];

export const PinType = {
  power: "power",
  ground: "ground",
  signal: "signal",
  analog: "analog",
  digital: "digital",
  bidirectional: "bidirectional",
} as const;

export type PinDirection = (typeof PinDirection)[keyof typeof PinDirection];

export const PinDirection = {
  in: "in",
  out: "out",
  passive: "passive",
  power: "power",
  ground: "ground",
  open_collector: "open_collector",
} as const;

export interface Pin {
  name: string;
  type: PinType;
  direction: PinDirection;
  maxVoltage?: number;
  driveVoltage?: number;
  pinNumber: number;
  net?: string;
}

export type NetlistComponentProperties = { [key: string]: string };

export interface NetlistComponent {
  id: string;
  name: string;
  type: string;
  category: NetlistComponentCategory;
  voltageLevel?: number;
  pins: Pin[];
  properties?: NetlistComponentProperties;
}

export interface NetlistConnection {
  from: string;
  fromPin: string;
  to: string;
  toPin: string;
  net: string;
}

export type NetType = (typeof NetType)[keyof typeof NetType];

export const NetType = {
  power: "power",
  ground: "ground",
  signal: "signal",
} as const;

export interface Net {
  name: string;
  type: NetType;
  voltage?: number;
}

export interface Netlist {
  components: NetlistComponent[];
  connections: NetlistConnection[];
  nets: Net[];
}

export interface CompileResult {
  success: boolean;
  errors: CompileError[];
  warnings: CompileWarning[];
  netlist?: Netlist;
}

export interface Example {
  name: string;
  description: string;
  source: string;
}

export interface ExampleList {
  examples: Example[];
}

export type ComponentInfoCategory =
  (typeof ComponentInfoCategory)[keyof typeof ComponentInfoCategory];

export const ComponentInfoCategory = {
  passive: "passive",
  active_discrete: "active_discrete",
  active_ic: "active_ic",
  sensor: "sensor",
  module: "module",
  power: "power",
} as const;

export interface ComponentInfo {
  type: string;
  category: ComponentInfoCategory;
  description: string;
  aliases?: string[];
  voltageLevel?: number;
  pins: Pin[];
}

export interface ComponentLibrary {
  components: ComponentInfo[];
}

export type ExportRequestFormat =
  (typeof ExportRequestFormat)[keyof typeof ExportRequestFormat];

export const ExportRequestFormat = {
  kicad: "kicad",
  proteus: "proteus",
  spice: "spice",
} as const;

export interface ExportRequest {
  netlist: Netlist;
  format?: ExportRequestFormat;
  title?: string;
}

export interface ExportResult {
  /** File content of the exported netlist */
  content: string;
  filename: string;
  format: string;
}

export interface LlmAnalyzeRequest {
  source: string;
  netlist?: Netlist;
}

export type RiskItemSeverity =
  (typeof RiskItemSeverity)[keyof typeof RiskItemSeverity];

export const RiskItemSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface RiskItem {
  severity: RiskItemSeverity;
  description: string;
  component?: string;
}

export interface LlmAnalyzeResult {
  /**
   * @minimum 0
   * @maximum 100
   */
  safetyScore: number;
  analysis: string;
  risks: RiskItem[];
  suggestions: string[];
  model: string;
}