# lib/api-zod/src/generated/types/compileError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { CompileErrorSeverity } from "./compileErrorSeverity";

export interface CompileError {
  line: number;
  column: number;
  message: string;
  errorCode: string;
  severity: CompileErrorSeverity;
}