# lib/api-zod/src/generated/types/compileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */

export interface CompileRequest {
  /** Rust-inspired circuit definition source code */
  source: string;
}