# lib/api-zod/src/generated/types/componentLibrary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { ComponentInfo } from "./componentInfo";

export interface ComponentLibrary {
  components: ComponentInfo[];
}