# lib/api-zod/src/generated/types/exportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { ExportRequestFormat } from "./exportRequestFormat";
import type { Netlist } from "./netlist";

export interface ExportRequest {
  netlist: Netlist;
  format?: ExportRequestFormat;
  title?: string;
}