# lib/api-zod/src/generated/types/exportResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */

export interface ExportResult {
  /** File content of the exported netlist */
  content: string;
  filename: string;
  format: string;
}