# lib/api-zod/src/generated/types/llmAnalyzeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { RiskItem } from "./riskItem";

export interface LlmAnalyzeResult {
  /**
   * @minimum 0
   * @maximum 100
   */
  safetyScore: number;
  analysis: string;
  risks: RiskItem[];
  suggestions: string[];
  model: string;
}