# lib/api-zod/src/generated/types/net.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { NetType } from "./netType";

export interface Net {
  name: string;
  type: NetType;
  voltage?: number;
}