# lib/api-zod/src/generated/types/netlist.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { Net } from "./net";
import type { NetlistComponent } from "./netlistComponent";
import type { NetlistConnection } from "./netlistConnection";

export interface Netlist {
  components: NetlistComponent[];
  connections: NetlistConnection[];
  nets: Net[];
}