# lib/api-zod/src/generated/types/netlistConnection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */

export interface NetlistConnection {
  from: string;
  fromPin: string;
  to: string;
  toPin: string;
  net: string;
}