# lib/api-zod/src/generated/types/riskItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Strict Circuit Compiler API
 * OpenAPI spec version: 0.2.0
 */
import type { RiskItemSeverity } from "./riskItemSeverity";

export interface RiskItem {
  severity: RiskItemSeverity;
  description: string;
  component?: string;
}